updating model

+++ src/Projector/src/main/java/com/itechart/kalenik/dao/repository/EmployeeRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface EmployeeRepository extends JpaRepository<Employee, Integer> {
+public interface EmployeeRepository extends JpaRepository<Employee, Long> {
     Employee findByUsername(String username);
-    Employee findById(Integer id);
+    Employee findById(Long id);
 }
