+++ src/Projector/src/main/java/com/itechart/kalenik/dao/repository/SubtaskRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface SubtaskRepository extends JpaRepository<Subtask, Integer> {
+public interface SubtaskRepository extends JpaRepository<Subtask, Long> {
 }
